Adds sticky sidenav
* Adds sticky sidenav element
* Sidenav links are highlighted based on position in page
* Clicking the links gives a fast, smooth scroll to the right place
(offset used to position in line with top of sidenav menu)

A couple of CSS changes for this and surrounding div on the sidenav, I
may be able to remove that.

And of course, the JS needs moving to a file.

Also, I can get rid of the toTop and replace with a scrollTo
implementation.